move base model to common
