+++ busy_beaver/common/models.py
